Alex's fixes to rlib plus bug fixes for region-persist-test.c
git-svn-id: https://banshee.svn.sourceforge.net/svnroot/banshee/trunk@170 c3c2f706-f20d-0410-965c-b05b340ef3e4

+++ banshee/rlib/regions.c
@@ -81,7 +81,7 @@ struct array_hdr
 #define ALIGN(x, n) (((x) + ((n) - 1)) & ~((n) - 1))
 #define PALIGN(x, n) ((void *)ALIGN((__rcintptr)(x), n))
 #ifdef __GNUC__
-#define RALIGNMENT __alignof(double)
+#define RALIGNMENT __alignof(unsigned long) /* double would be safer */
 #define ALIGNMENT_HDR __alignof(adjust_fn *)
 #define PTRALIGNMENT __alignof(void *)
 #define ALIGNMENT_LONG __alignof(unsigned long)
